+++ src/js/@shared/utils/index.js
@@ -1,3 +1,4 @@
 export { $, parseToElements } from './dom';
 export { getFromSessionStorage, setToSessionStorage, removeFromSessionStorage } from './sessionStorage';
 export { encrypt } from './encryption';
+export { request } from './api';
